Changes to swap scripts that checks if containers exist

+++ HW4/swap2to1.sh
@@ -3,6 +3,10 @@
 
 # time1=$(docker inspect ecs189_web2_1)
 # echo $time1 > sample.json
+if ! [ "$(docker ps | grep ecs189_web2_1)" ]; then
+	echo "docker image ecs189_web2 does not exist"
+	exit
+fi
 
 docker kill ecs189_web2_1
 sleep 5
@@ -13,10 +17,10 @@ docker network rm $(docker network ls | grep "bridge" | awk '/ / { print $1 }')
 sleep 5
 
 docker run --network ecs189_default -d --name ecs189_web1_1 --link ecs189_proxy_1 -p 8080 activity
-sleep 10
+sleep 5
 
 docker exec ecs189_proxy_1 /bin/bash /bin/swap1.sh
-sleep 20
+sleep 10
 
 echo "done"
